refactor: 카드폼 작성 관련 Input들을 각각의 디렉토리로 분리
