+++ src/components/Container/index.stories.js
@@ -47,10 +47,10 @@ ExpirationDateInputContainer.args = {
 
 export const OwnerNameInputContainer = Template.bind({});
 OwnerNameInputContainer.args = {
-  className: 'CardInfoForm__Input__Filler--filled CardOwnerInput__Filler',
+  className: 'CardInfoForm__Input__Filler--filled OwnerNameInput__Filler',
   children: (
     <Input
-      className="CardOwnerInput__Field"
+      className="OwnerNameInput__Field"
       placeholder="카드에 표시된 이름과 동일하게 입력하세요."
       value="SUN"
     />
